backend/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
dj-database-url==1.2.0
Django==4.1.4
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
whitenoise==6.2.0
